refactor(types): move DocumentRange/RelativeRange to shared/positions/types
Position-system types belong in the position module, not the general
shared types file. Re-exported from shared/types.ts so existing
importers continue to work. Direct importers updated to canonical path.

Closes #71
