+++ src/shared/positions/index.ts
@@ -1,4 +1,6 @@
 export type {
+  DocumentRange,
+  RelativeRange,
   RangeValidation,
   AnchoredRangeResult,
   ElementPosition,
